refactor: separate each new layout component to its own file
