+++ public/app/pages/PageContentWrapper.tsx
@@ -1,0 +1,5 @@
+import React, { ReactNode } from 'react';
+
+export function PageContentWrapper({ children }: { children: ReactNode }) {
+  return <div className="main-wrapper">{children}</div>;
+}
